+++ Joveler.DynLoader.Tests/GhostErrorTests.cs
@@ -1,5 +1,5 @@
 ﻿/*
-    Copyright (C) 2019-2021 Hajin Jang
+    Copyright (C) 2019-2023 Hajin Jang
     Licensed under MIT License.
  
     MIT License
@@ -59,7 +59,7 @@ namespace Joveler.DynLoader.Tests
             try
             {
                 GhostFunction lib = new GhostFunction();
-                lib.LoadLibrary(TestSetup.PackagedZLibPath);
+                lib.LoadLibrary(TestSetup.PackagedZLibPathStdcall);
             }
             catch (EntryPointNotFoundException e)
             {